+++ include/rudiments/ucs2stringbuffer.h
@@ -63,7 +63,7 @@ class RUDIMENTS_DLLSPEC ucs2stringbuffer : public bytebuffer {
 
 		/** Returns the position in the internal buffer at which
 		 *  the next write will occur. */
-		size_t	getPosition();
+		off64_t	getPosition();
 
 		/** Writes "string" to the ucs2stringbuffer at the current
 		 *  position and increments the current position to the next
